Long -> int for clan ids
Just learned that clan ids of all regions fits within int32 so here we go

+++ Arona/ApiModels/LadderStructure.cs
@@ -4,7 +4,7 @@ namespace Arona.ApiModels;
 
 internal class LadderStructure
 {
-    [JsonPropertyName("id")] public required long Id { get; init; }
+    [JsonPropertyName("id")] public required int Id { get; init; }
     [JsonPropertyName("tag")] public required string Tag { get; init; }
     [JsonPropertyName("name")] public required string Name { get; init; }
     [JsonPropertyName("rank")] public required int Rank { get; init; }
